fix: move integration fixtures to top-level conftest
Pytest 9 disallows `pytest_plugins` in subdirectory conftests, and the
workaround (declaring it in tests/conftest.py) caused double-registration
because auto-discovery loaded tests/integration/conftest.py too. Cleanest
fix: move the four session fixtures (real_connector, test_group,
tmp_contact, integration_env) and their helpers up to tests/conftest.py
so both tests/integration/ and tests/benchmarks/ inherit them via normal
parent-directory walking. No pytest_plugins line needed.

tests/integration/conftest.py becomes a docstring-only stub documenting
where the fixtures moved to and the original lifecycle.

Locally: 486 unit tests pass, 14 benchmarks pass, no regressions.
